+++ transpiler/__tests__/end-to-end/mocks/symbol-table/already-declared/index.ts
@@ -1,0 +1,25 @@
+import { FileUtil } from '../../../../../src/utils/file.js';
+import { SymbolTableErrorTestCase } from '../symbol-table.js';
+
+export const SYMBOL_TABLE_ALREADY_DECLARED_TEST_CASES: SymbolTableErrorTestCase[] = [
+  {
+    description: 'Should return error that const declaration is already defined',
+    inputCore: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/symbol-table/already-declared/already-declared-const.bl',
+    ),
+    inputSetup: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/semantic-errors/setup.bl',
+    ),
+    errorMessages: ['Identifier accountId is already defined.'],
+  },
+  {
+    description: 'Should return error that variable declaration is already defined',
+    inputCore: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/symbol-table/already-declared/already-declared-variable.bl',
+    ),
+    inputSetup: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/semantic-errors/setup.bl',
+    ),
+    errorMessages: ['Identifier accountId is already defined.'],
+  },
+];
