+++ transpiler/__tests__/end-to-end/mocks/symbol-table/command-handler-static-method.bl
@@ -18,10 +18,8 @@ CommandHandler WithdrawMoneyCommandHandler (accountRepo: AccountWriteRepoPort) {
     execute(command: WithdrawMoneyCommand): (OK(void), Errors(
        ApplicationErrors.AccountNotFound
        )) {
-        
-        const fromPrimitivesInput = FromPrimitivesInput({balance: command.amount, id: command.accountId});
-       
-        const accountEntity = AccountEntity.fromPrimitives(fromPrimitivesInput);
+               
+        const accountEntity = AccountEntity.withdrawAmount(command.amount).ifError();
    }
  }
 
@@ -29,7 +27,7 @@ Root Entity AccountEntity {
     static create(props: AccountProps): (OK(AccountEntity), Errors())  {
     }
 
-    public withdrawAmount(amount: int32): (OK(void), Errors(DomainErrors.InvalidMonetaryValueError)) {
+    public static withdrawAmount(amount: int32): (OK(void), Errors(DomainErrors.InvalidMonetaryValueError)) {
     }
 
 }
