+++ transpiler/__tests__/end-to-end/mocks/symbol-table/constant-reassignment/index.ts
@@ -1,0 +1,15 @@
+import { FileUtil } from '../../../../../src/utils/file.js';
+import { SymbolTableErrorTestCase } from '../symbol-table.js';
+
+export const SYMBOL_TABLE_CONSTANT_REASSIGNMENT_TEST_CASES: SymbolTableErrorTestCase[] = [
+  {
+    description: 'Should return error that const declaration cannot be reassigned',
+    inputCore: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/symbol-table/constant-reassignment/const-reassignment.bl',
+    ),
+    inputSetup: FileUtil.readFileString(
+      'transpiler/__tests__/end-to-end/mocks/semantic-errors/setup.bl',
+    ),
+    errorMessages: ['Cannot reassign constant variable accountId.'],
+  },
+];
