+++ transpiler/src/ast/core/intermediate-ast/nodes/Entity/EntityDeclarationNode.ts
@@ -5,7 +5,6 @@ import { DomainCreateNode } from '../Domain/DomainCreateNode.js';
 import { TNodeMetadata } from '../IntermediateASTNode.js';
 import { EntityIdentifierNode } from './EntityIdentifierNode.js';
 import { EntityValuesNode } from './EntityValuesNode.js';
-import { MethodCallSymbolEntry } from '../../../../../semantic-analysis/type-inference/SymbolEntry.js';
 
 export class EntityDeclarationNode extends ClassTypeNode {
   private static classType = ClassTypes.Entity;
@@ -39,17 +38,8 @@ export class EntityDeclarationNode extends ClassTypeNode {
     return entityValues;
   }
 
-  public addFromPrimitivesToSymbolTable(symbolTableManager: SymbolTableManager): void {
-    const initialSymbolTable = symbolTableManager.getSymbolTable();
-    initialSymbolTable.insert(
-      SymbolTableManager.FROM_PRIMITIVES + '()',
-      new MethodCallSymbolEntry(this.getIdentifier().getIdentifierName()),
-    );
-  }
-
   addToSymbolTable(symbolTableManager: SymbolTableManager): void {
     symbolTableManager.addClassTypeThis(this.getIdentifier().getIdentifierName());
-    this.addFromPrimitivesToSymbolTable(symbolTableManager);
     const entityValues = this.getEntityValues();
     entityValues.addToSymbolTable(symbolTableManager);
   }
