+++ transpiler/src/ast/core/intermediate-ast/nodes/RootEntity/RootEntityDeclarationNode.ts
@@ -7,7 +7,6 @@ import { TNodeMetadata } from '../IntermediateASTNode.js';
 import { EntityValuesNode } from '../Entity/EntityValuesNode.js';
 import { PublicMethodDeclarationNode } from '../methods/PublicMethodDeclarationNode.js';
 import { SymbolTableManager } from '../../../../../semantic-analysis/type-inference/SymbolTableManager.js';
-import { MethodCallSymbolEntry } from '../../../../../semantic-analysis/type-inference/SymbolEntry.js';
 
 //TODO fix this and entity to be extendable from a base class
 export class RootEntityDeclarationNode extends ClassTypeNode {
@@ -51,17 +50,8 @@ export class RootEntityDeclarationNode extends ClassTypeNode {
     return entityValues;
   }
 
-  public addFromPrimitivesToSymbolTable(symbolTableManager: SymbolTableManager): void {
-    const initialSymbolTable = symbolTableManager.getSymbolTable();
-    initialSymbolTable.insert(
-      SymbolTableManager.FROM_PRIMITIVES + '()',
-      new MethodCallSymbolEntry(this.getIdentifier().getIdentifierName()),
-    );
-  }
-
   addToSymbolTable(symbolTableManager: SymbolTableManager): void {
     symbolTableManager.addClassTypeThis(this.getIdentifier().getIdentifierName());
-    this.addFromPrimitivesToSymbolTable(symbolTableManager);
     const entityValues = this.getEntityValues();
     entityValues.addToSymbolTable(symbolTableManager);
   }
