+++ transpiler/src/semantic-analysis/type-inference/SymbolTableManager.ts
@@ -22,7 +22,6 @@ export class SymbolTableManager {
     HANDLE: 'handle',
     IF_ERROR: 'ifError',
   };
-  public static readonly FROM_PRIMITIVES = 'fromPrimitives';
   private symbolTable: SymbolTable;
   private intermediateASTTree: IntermediateASTTree;
   private boundedContexts: TBoundedContexts;
